presence [nfc]: Consolidate "utils" file into model
